fix: css change and form
